+++ project_template/game/assets/__init__.py
