--- project_template/rename/setup.py
@@ -1,3 +1,0 @@
-class Setup():
-    def __init__(self):
-
